VZ-5858 move function
